requirements.txt
Flask==2.3.2
opencv-python==4.8.0.76
numpy==1.24.5
ultralytics==8.0.20
transformers==4.31.0
Pillow==10.0.1
paddleocr==2.6.1.2
paddlepaddle==2.5.2
inference-sdk==1.3.2
python-dotenv==1.0.0
threading2==0.2.6
queue==0.6.2
